+++ contracts/utils/cryptography/signers/MultiSignerERC7913.sol
@@ -3,9 +3,8 @@
 pragma solidity ^0.8.27;
 
 import {AbstractSigner} from "./AbstractSigner.sol";
-import {ERC7913Utils} from "../ERC7913Utils.sol";
+import {SignatureChecker} from "../SignatureChecker.sol";
 import {EnumerableSet} from "../../structs/EnumerableSet.sol";
-import {Calldata} from "../../../utils/Calldata.sol";
 import {SafeCast} from "../../../utils/math/SafeCast.sol";
 
 /**
@@ -47,17 +46,17 @@ import {SafeCast} from "../../../utils/math/SafeCast.sol";
  */
 abstract contract MultiSignerERC7913 is AbstractSigner {
     using EnumerableSet for EnumerableSet.BytesSet;
-    using ERC7913Utils for *;
+    using SignatureChecker for *;
     using SafeCast for uint256;
 
-    EnumerableSet.BytesSet private _signersSet;
+    EnumerableSet.BytesSet private _signers;
     uint128 private _threshold;
 
-    /// @dev Emitted when signers are added.
-    event ERC7913SignersAdded(bytes[] indexed signers);
+    /// @dev Emitted when a signer is added.
+    event ERC7913SignerAdded(bytes indexed signers);
 
-    /// @dev Emitted when signers are removed.
-    event ERC7913SignersRemoved(bytes[] indexed signers);
+    /// @dev Emitted when a signers is removed.
+    event ERC7913SignerRemoved(bytes indexed signers);
 
     /// @dev Emitted when the threshold is updated.
     event ERC7913ThresholdSet(uint256 threshold);
@@ -82,22 +81,17 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
      * if the signers set grows too large.
      */
     function signers() public view virtual returns (bytes[] memory) {
-        return _signers().values();
+        return _signers.values();
     }
 
     /// @dev Returns whether the `signer` is an authorized signer.
     function isSigner(bytes memory signer) public view virtual returns (bool) {
-        return _signers().contains(signer);
+        return _signers.contains(signer);
     }
 
     /// @dev Returns the minimum number of signers required to approve a multisignature operation.
     function threshold() public view virtual returns (uint256) {
         return _threshold;
-    }
-
-    /// @dev Returns the set of authorized signers.
-    function _signers() internal view virtual returns (EnumerableSet.BytesSet storage) {
-        return _signersSet;
     }
 
     /**
@@ -110,13 +104,12 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
      * * Each of `newSigners` must not be authorized. See {isSigner}. Reverts with {MultiSignerERC7913AlreadyExists} if so.
      */
     function _addSigners(bytes[] memory newSigners) internal virtual {
-        uint256 newSignersLength = newSigners.length;
-        for (uint256 i = 0; i < newSignersLength; i++) {
+        for (uint256 i = 0; i < newSigners.length; ++i) {
             bytes memory signer = newSigners[i];
             require(signer.length >= 20, MultiSignerERC7913InvalidSigner(signer));
-            require(_signers().add(signer), MultiSignerERC7913AlreadyExists(signer));
+            require(_signers.add(signer), MultiSignerERC7913AlreadyExists(signer));
+            emit ERC7913SignerAdded(signer);
         }
-        emit ERC7913SignersAdded(newSigners);
     }
 
     /**
@@ -128,13 +121,12 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
      * * See {_validateReachableThreshold} for the threshold validation.
      */
     function _removeSigners(bytes[] memory oldSigners) internal virtual {
-        uint256 oldSignersLength = oldSigners.length;
-        for (uint256 i = 0; i < oldSignersLength; i++) {
+        for (uint256 i = 0; i < oldSigners.length; ++i) {
             bytes memory signer = oldSigners[i];
-            require(_signers().remove(signer), MultiSignerERC7913NonexistentSigner(signer));
+            require(_signers.remove(signer), MultiSignerERC7913NonexistentSigner(signer));
+            emit ERC7913SignerRemoved(signer);
         }
         _validateReachableThreshold();
-        emit ERC7913SignersRemoved(oldSigners);
     }
 
     /**
@@ -159,7 +151,7 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
      * * The {signers}'s length must be `>=` to the {threshold}. Throws {MultiSignerERC7913UnreachableThreshold} if not.
      */
     function _validateReachableThreshold() internal view virtual {
-        uint256 totalSigners = _signers().length();
+        uint256 totalSigners = _signers.length();
         uint256 currentThreshold = threshold();
         require(
             totalSigners >= currentThreshold,
@@ -206,7 +198,6 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
     ) internal view virtual override returns (bool) {
         if (signature.length == 0) return false; // For ERC-7739 compatibility
         (bytes[] memory signingSigners, bytes[] memory signatures) = abi.decode(signature, (bytes[], bytes[]));
-        if (signingSigners.length != signatures.length) return false;
         return _validateThreshold(signingSigners) && _validateSignatures(hash, signingSigners, signatures);
     }
 
@@ -227,13 +218,12 @@ abstract contract MultiSignerERC7913 is AbstractSigner {
         bytes[] memory signingSigners,
         bytes[] memory signatures
     ) internal view virtual returns (bool valid) {
-        uint256 signersLength = signingSigners.length;
-        for (uint256 i = 0; i < signersLength; i++) {
+        for (uint256 i = 0; i < signingSigners.length; ++i) {
             if (!isSigner(signingSigners[i])) {
                 return false;
             }
         }
-        return hash.areValidSignaturesNow(signingSigners, signatures);
+        return hash.areValidERC7913SignaturesNow(signingSigners, signatures);
     }
 
     /**
