+++ docs/modules/ROOT/pages/utilities.adoc
@@ -97,30 +97,79 @@ function _verify(
 
 IMPORTANT: Always use keys of at least 2048 bits. Additionally, be aware that PKCS#1 v1.5 allows for replayability due to the possibility of arbitrary optional parameters. To prevent replay attacks, consider including an onchain nonce or unique identifier in the message.
 
-=== ERC-7913 Signature Verifiers
+=== Signature Verification
 
-ERC-7913 extends the concept of signature verification to support keys that don't have their own Ethereum address. This is particularly useful for integrating non-Ethereum cryptographic curves, hardware devices, or other identity systems into smart accounts.
+The xref:api:utils.adoc#SignatureChecker[`SignatureChecker`] library provides a unified interface for verifying signatures from different sources. It seamlessly supports:
 
-The standard defines a verifier interface that can be implemented to support different types of keys. A signer is represented as a `bytes` object that concatenates a verifier address and a key: `verifier || key`.
+* ECDSA signatures from externally owned accounts (EOAs)
+* ERC-1271 signatures from smart contract wallets like Argent and Safe Wallet
+* ERC-7913 signatures from keys that don't have their own Ethereum address
 
-xref:api:utils.adoc#ERC7913Utils[`ERC7913Utils`] provides functions for verifying signatures using ERC-7913 compatible verifiers:
+This allows developers to write signature verification code once and have it work across all these different signature types.
+
+==== Basic Signature Verification
+
+For standard signature verification that supports both EOAs and ERC-1271 contracts:
 
 [source,solidity]
 ----
-using ERC7913Utils for bytes;
+using SignatureChecker for address;
 
-function _verify(bytes memory signer, bytes32 hash, bytes memory signature) internal view returns (bool) {
-    return signer.isValidSignatureNow(hash, signature);
+function _verifySignature(address signer, bytes32 hash, bytes memory signature) internal view returns (bool) {
+    return SignatureChecker.isValidSignatureNow(signer, hash, signature);
+}
+----
+
+The library automatically detects whether the signer is an EOA or a contract and uses the appropriate verification method.
+
+==== ERC-1271 Contract Signatures
+
+For smart contract wallets that implement ERC-1271, you can explicitly use:
+
+[source,solidity]
+----
+function _verifyContractSignature(address signer, bytes32 hash, bytes memory signature) internal view returns (bool) {
+    return SignatureChecker.isValidERC1271SignatureNow(signer, hash, signature);
+}
+----
+
+==== ERC-7913 Extended Signatures
+
+ERC-7913 extends signature verification to support keys that don't have their own Ethereum address. This is useful for integrating non-Ethereum cryptographic curves, hardware devices, or other identity systems.
+
+A signer is represented as a `bytes` object that concatenates a verifier address and a key: `verifier || key`.
+
+[source,solidity]
+----
+function _verifyERC7913Signature(bytes memory signer, bytes32 hash, bytes memory signature) internal view returns (bool) {
+    return SignatureChecker.isValidERC7913SignatureNow(signer, hash, signature);
 }
 ----
 
 The verification process works as follows:
 
 * If `signer.length < 20`: verification fails
-* If `signer.length == 20`: verification is done using xref:api:utils#SignatureChecker[SignatureChecker]
-* Otherwise: verification is done using an ERC-7913 verifier.
+* If `signer.length == 20`: verification is done using standard signature checking
+* Otherwise: verification is done using an ERC-7913 verifier
 
-This allows for backward compatibility with EOAs and ERC-1271 contracts while supporting new types of keys.
+==== Batch Verification
+
+For verifying multiple ERC-7913 signatures at once:
+
+[source,solidity]
+----
+function _verifyMultipleSignatures(
+    bytes32 hash,
+    bytes[] memory signers,
+    bytes[] memory signatures
+) internal view returns (bool) {
+    return SignatureChecker.areValidERC7913SignaturesNow(hash, signers, signatures);
+}
+----
+
+The signers must be ordered by their `keccak256` hash to ensure no duplicates and optimize verification.
+
+This unified approach allows smart contracts to accept signatures from any supported source without needing to implement different verification logic for each type.
 
 === Verifying Merkle Proofs
